Modularized the command line interface
