+++ src/lib.rs
@@ -2,6 +2,9 @@
 #![allow(clippy::missing_errors_doc)]
 // TODO
 #![allow(clippy::cast_possible_truncation)]
+#![allow(clippy::cast_precision_loss)]
+#![allow(clippy::map_unwrap_or)]
 
+pub mod cli;
 pub mod interpreter;
 pub mod program;
